120 -> 100 col
moving to 100 because 120 is too wide for most of the work I do. I avoid
more than 3 levels of indentation anyways as a strict coding convention
and with 4spaces of tab, with 100 I'd get atleast 88 chars for myself
anyways.

I would have still stuck to 80 but some API calls require string input,
which I can dodge with declaring local vars but ehhhh might as well update
my preference with time at this point.

120 is still quite too extreme for my case though. My monitor isn't even
wide enough to support 240+ chars in a line if im working on nvim with
two panels + relativenumber + gitgutter. 90 is the limit for me with
'cause of that, which lies conveniently between 80 and 100.

100 doesn't mean I have to change my coding habits or try to push to 100
with every line. I would still follow my old convention and try to fit
everything within 80. This change only means that I would not fuzz about
my code if it exceeds 80 (unless it also exceeds 100).

Anyways this is a git commit message. prolly never going to be read by
anyone ever, including future me. still i felt like justifying my change
from 80 to 120 to 100. unless i get a wider monitor or am part of a team
with different convention, 100 is what im going to stick to from this
point.